changed validation token to validate api tokens not backend created token
