+++ web/client/src/components/index.js
@@ -2,6 +2,6 @@ import Head from "./Head";
 import TokenForm from "./form/TokenForm";
 import TokenType from "./form/TokenType";
 import TokenScopes from "./form/TokenScopes";
-import UseToken from "./TokenValidation/UseToken";
+import ValidateToken from "./validation/TokenValidation";
 
-export { Head, TokenForm, TokenScopes, TokenType, UseToken };
+export { Head, TokenForm, TokenScopes, TokenType, ValidateToken };
